refactor: owner row
