+++ client/residential_management/lib/owners/widgets/owner_row/widgets/owner_name.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:residential_management/functions/name_formatter.dart';
+
+class OwnerName extends StatelessWidget {
+  const OwnerName({required this.firstName, required this.lastName, super.key});
+
+  final String firstName;
+  final String lastName;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(left: 16, top: 16),
+      child: Text(
+        nameFormatter(firstName, lastName),
+        style: const TextStyle(
+          fontSize: 18,
+          fontWeight: FontWeight.w600,
+          height: 1,
+        ),
+      ),
+    );
+  }
+}
